add lib.rs and move main.rs code
